separate out section routes
