+++ app/routes.js
@@ -6,16 +6,10 @@ import config from './config';
 import { logger } from './logger';
 import { withCatch, getHealthCheckDependencies, extractAccessToken } from './helpers/routerHelper';
 import { getDashboardContext } from './pages/dashboard/controller';
-import { getDescriptionContext, getDescriptionErrorContext, postOrPutDescription } from './pages/sections/description/controller';
-import {
-  getSupplierSearchPageContext, validateSupplierSearchForm, getSupplierSearchPageErrorContext,
-} from './pages/sections/supplier/search/controller';
-import { findSuppliers, getSupplierSelectPageContext } from './pages/sections/supplier/select/controller';
+import { sectionRoutes } from './pages/sections/routes';
 import includesContext from './includes/manifest.json';
 import { getTaskListPageContext } from './pages/task-list/controller';
-import {
-  getCallOffOrderingPartyContext, getCallOffOrderingPartyErrorContext, putCallOffOrderingParty,
-} from './pages/sections/call-off-ordering-party/controller';
+
 
 const addContext = ({ context, user, csrfToken }) => ({
   ...context,
@@ -36,6 +30,7 @@ export const routes = (authProvider) => {
   });
 
   router.get('/', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
+    logger.info('redirecting to organisation orders page');
     res.redirect(`${config.baseUrl}/organisation`);
   }));
 
@@ -46,6 +41,7 @@ export const routes = (authProvider) => {
       orgId: req.user.primaryOrganisationId,
       orgName: req.user.primaryOrganisationName,
     });
+    logger.info('navigating to organisation orders page');
     res.render('pages/dashboard/template.njk', addContext({ context, user: req.user }));
   }));
 
@@ -53,115 +49,11 @@ export const routes = (authProvider) => {
     const accessToken = extractAccessToken({ req, tokenType: 'access' });
     const { orderId } = req.params;
     const context = await getTaskListPageContext({ accessToken, orderId });
+    logger.info(`navigating to order ${orderId} task list page`);
     res.render('pages/task-list/template.njk', addContext({ context, user: req.user }));
   }));
 
-  router.get('/organisation/:orderId/description', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const context = await getDescriptionContext({ orderId, accessToken: extractAccessToken({ req, tokenType: 'access' }) });
-    res.render('pages/sections/description/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.post('/organisation/:orderId/description', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-    const response = await postOrPutDescription({
-      orgId: req.user.primaryOrganisationId,
-      orderId,
-      data: req.body,
-      accessToken,
-    });
-
-    if (response.success) return res.redirect(`${config.baseUrl}/organisation/${response.orderId}`);
-
-    const context = await getDescriptionErrorContext({
-      validationErrors: response.errors,
-      orderId,
-      data: req.body,
-    });
-    return res.render('pages/sections/description/template', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.get('/organisation/:orderId/call-off-ordering-party', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const orgId = req.user.primaryOrganisationId;
-    const context = await getCallOffOrderingPartyContext({ orderId, orgId, accessToken: extractAccessToken({ req, tokenType: 'access' }) });
-    res.render('pages/sections/call-off-ordering-party/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.post('/organisation/:orderId/call-off-ordering-party', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const response = await putCallOffOrderingParty({
-      orgId: req.user.primaryOrganisationId,
-      orderId,
-      data: req.body,
-      accessToken: extractAccessToken({ req, tokenType: 'access' }),
-    });
-    if (response.success) return res.redirect(`${config.baseUrl}/organisation/${orderId}`);
-
-    const context = await getCallOffOrderingPartyErrorContext({
-      validationErrors: response.errors,
-      orderId,
-      data: req.body,
-    });
-    return res.render('pages/sections/call-off-ordering-party//template', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.get('/organisation/:orderId/supplier', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    res.redirect(`${config.baseUrl}/organisation/${orderId}/supplier/search`);
-  }));
-
-  router.get('/organisation/:orderId/supplier/search', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const context = await getSupplierSearchPageContext({ orderId });
-    res.render('pages/sections/supplier/search/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.post('/organisation/:orderId/supplier/search', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-
-    const response = validateSupplierSearchForm({ data: req.body });
-
-    if (response.success) {
-      return res.redirect(`${config.baseUrl}/organisation/${orderId}/supplier/search?name=${req.body.supplierName}`);
-    }
-
-    const context = await getSupplierSearchPageErrorContext({
-      orderId,
-      validationErrors: response.errors,
-    });
-
-    return res.render('pages/sections/supplier/search/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.get('/organisation/:orderId/supplier/search/select', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const { name } = req.query;
-
-    if (name) {
-      const accessToken = extractAccessToken({ req, tokenType: 'access' });
-
-      const suppliersFound = await findSuppliers({
-        name, accessToken,
-      });
-
-      if (suppliersFound.length > 0) {
-        const context = getSupplierSelectPageContext({ orderId, suppliers: suppliersFound });
-        return res.render('pages/sections/supplier/select/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-      }
-
-      throw new ErrorContext({
-        status: 404,
-        title: 'No Supplier found',
-        description: "There are no suppliers that match the search terms you've provided. Try searching again.",
-        backLinkText: 'Go back to search',
-        backLinkHref: `${config.baseUrl}/organisation/${orderId}/supplier/search`,
-      });
-    }
-
-    return res.redirect(`${config.baseUrl}/organisation/${orderId}/supplier/search`);
-  }));
+  router.use('/organisation/:orderId', sectionRoutes(authProvider, addContext));
 
   router.get('*', (req) => {
     throw new ErrorContext({
