+++ app/routes.test.js
@@ -2,37 +2,17 @@ import request from 'supertest';
 import {
   FakeAuthProvider,
   testAuthorisedGetPathForUnauthenticatedUser,
-  testPostPathWithoutCsrf,
-  testAuthorisedPostPathForUnauthenticatedUser,
-  testAuthorisedPostPathForUnauthorisedUsers,
   testAuthorisedGetPathForUnauthorisedUser,
-  getCsrfTokenFromGet,
 } from 'buying-catalogue-library';
 import { App } from './app';
 import { routes } from './routes';
 import { baseUrl } from './config';
 import * as dashboardController from './pages/dashboard/controller';
 import * as taskListController from './pages/task-list/controller';
-import * as descriptionController from './pages/sections/description/controller';
-import * as orderingPartyController from './pages/sections/call-off-ordering-party/controller';
-import * as supplierSearchController from './pages/sections/supplier/search/controller';
-import * as supplierSelectController from './pages/sections/supplier/select/controller';
 
 jest.mock('./logger');
 
 dashboardController.getDashboardContext = jest.fn()
-  .mockResolvedValue({});
-
-descriptionController.getDescriptionContext = jest.fn()
-  .mockResolvedValue({});
-
-descriptionController.postOrPutDescription = jest.fn()
-  .mockResolvedValue({});
-
-orderingPartyController.getCallOffOrderingPartyContext = jest.fn()
-  .mockResolvedValue({});
-
-orderingPartyController.putCallOffOrderingParty = jest.fn()
   .mockResolvedValue({});
 
 const mockLogoutMethod = jest.fn().mockImplementation(() => Promise.resolve({}));
@@ -151,429 +131,6 @@ describe('routes', () => {
         .then((res) => {
           expect(res.text.includes('data-test-id="order-id-page"')).toBeTruthy();
           expect(res.text.includes('data-test-id="error-title"')).toBeFalsy();
-        });
-    });
-  });
-
-  describe('GET /organisation/:orderId/description', () => {
-    const path = '/organisation/some-order-id/description';
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedGetPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()), pathToTest: path, expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedGetPathForUnauthorisedUser({
-        app: request(setUpFakeApp()),
-        pathToTest: path,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should return the correct status and text when the user is authorised', () => request(setUpFakeApp())
-      .get(path)
-      .set('Cookie', [mockAuthorisedCookie])
-      .expect(200)
-      .then((res) => {
-        expect(res.text.includes('data-test-id="description-page"')).toBeTruthy();
-        expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-      }));
-  });
-
-  describe('POST /organisation/neworder/description', () => {
-    const path = '/organisation/neworder/description';
-
-    afterEach(() => {
-      descriptionController.postOrPutDescription.mockReset();
-    });
-
-    it('should return 403 forbidden if no csrf token is available', () => (
-      testPostPathWithoutCsrf({
-        app: request(setUpFakeApp()), pathToTest: path, mockAuthorisedCookie,
-      })
-    ));
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedPostPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()),
-        csrfPagePath: path,
-        pathToTest: path,
-        mockAuthorisedCookie,
-        expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedPostPathForUnauthorisedUsers({
-        app: request(setUpFakeApp()),
-        csrfPagePath: path,
-        pathToTest: path,
-        mockAuthorisedCookie,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should return the correct status and text if response.success is true', async () => {
-      descriptionController.postOrPutDescription = jest.fn()
-        .mockImplementation(() => Promise.resolve({ success: true, orderId: 'order1' }));
-
-      const { cookies, csrfToken } = await getCsrfTokenFromGet({
-        app: request(setUpFakeApp()), csrfPagePath: path, mockAuthorisedCookie,
-      });
-
-      return request(setUpFakeApp())
-        .post(path)
-        .type('form')
-        .set('Cookie', [cookies, mockAuthorisedCookie])
-        .send({
-          description: 'a description of the order',
-          _csrf: csrfToken,
-        })
-        .expect(302)
-        .then((res) => {
-          expect(res.redirect).toEqual(true);
-          expect(res.headers.location).toEqual(`${baseUrl}/organisation/order1`);
-          expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-        });
-    });
-
-    it('should return the correct status and text if response.success is not true', async () => {
-      descriptionController.postOrPutDescription = jest.fn()
-        .mockImplementation(() => Promise.resolve({ success: false }));
-
-      descriptionController.getDescriptionErrorContext = jest.fn()
-        .mockImplementation(() => Promise.resolve({
-          errors: [{ text: 'Description too long', href: '#description' }],
-        }));
-
-      const { cookies, csrfToken } = await getCsrfTokenFromGet({
-        app: request(setUpFakeApp()), csrfPagePath: path, mockAuthorisedCookie,
-      });
-
-      return request(setUpFakeApp())
-        .post(path)
-        .type('form')
-        .set('Cookie', [cookies, mockAuthorisedCookie])
-        .send({ _csrf: csrfToken })
-        .expect(200)
-        .then((res) => {
-          expect(res.text.includes('data-test-id="description-page"')).toEqual(true);
-          expect(res.text.includes('data-test-id="error-summary"')).toEqual(true);
-          expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-          descriptionController.getDescriptionErrorContext.mockReset();
-        });
-    });
-  });
-
-  describe('GET /organisation/:orderId/call-off-ordering-party', () => {
-    const path = '/organisation/some-order-id/call-off-ordering-party';
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedGetPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()), pathToTest: path, expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedGetPathForUnauthorisedUser({
-        app: request(setUpFakeApp()),
-        pathToTest: path,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should return the correct status and text when the user is authorised', () => request(setUpFakeApp())
-      .get(path)
-      .set('Cookie', [mockAuthorisedCookie])
-      .expect(200)
-      .then((res) => {
-        expect(res.text.includes('data-test-id="call-off-ordering-party-page"')).toBeTruthy();
-        expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-      }));
-  });
-
-  describe('POST /organisation/neworder/call-off-ordering-party', () => {
-    const path = '/organisation/order-id/call-off-ordering-party';
-
-    afterEach(() => {
-      orderingPartyController.putCallOffOrderingParty.mockReset();
-    });
-
-    it('should return 403 forbidden if no csrf token is available', () => (
-      testPostPathWithoutCsrf({
-        app: request(setUpFakeApp()), pathToTest: path, mockAuthorisedCookie,
-      })
-    ));
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedPostPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()),
-        csrfPagePath: path,
-        pathToTest: path,
-        mockAuthorisedCookie,
-        expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedPostPathForUnauthorisedUsers({
-        app: request(setUpFakeApp()),
-        csrfPagePath: path,
-        pathToTest: path,
-        mockAuthorisedCookie,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should return the correct status and text if response.success is true', async () => {
-      orderingPartyController.putCallOffOrderingParty = jest.fn()
-        .mockImplementation(() => Promise.resolve({ success: true }));
-
-      const { cookies, csrfToken } = await getCsrfTokenFromGet({
-        app: request(setUpFakeApp()), csrfPagePath: path, mockAuthorisedCookie,
-      });
-
-      return request(setUpFakeApp())
-        .post(path)
-        .type('form')
-        .set('Cookie', [cookies, mockAuthorisedCookie])
-        .send({ _csrf: csrfToken })
-        .expect(302)
-        .then((res) => {
-          expect(res.redirect).toEqual(true);
-          expect(res.headers.location).toEqual(`${baseUrl}/organisation/order-id`);
-          expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-        });
-    });
-
-    it('should return the correct status and text if response.success is not true', async () => {
-      orderingPartyController.putCallOffOrderingParty = jest.fn()
-        .mockImplementation(() => Promise.resolve({ success: false }));
-
-      orderingPartyController.getCallOffOrderingPartyErrorContext = jest.fn()
-        .mockImplementation(() => Promise.resolve({
-          errors: [{ text: 'First name must be 100 characters or fewer', href: '#firstName' }],
-        }));
-
-      const { cookies, csrfToken } = await getCsrfTokenFromGet({
-        app: request(setUpFakeApp()), csrfPagePath: path, mockAuthorisedCookie,
-      });
-
-      return request(setUpFakeApp())
-        .post(path)
-        .type('form')
-        .set('Cookie', [cookies, mockAuthorisedCookie])
-        .send({ _csrf: csrfToken })
-        .expect(200)
-        .then((res) => {
-          expect(res.text.includes('data-test-id="call-off-ordering-party-page"')).toEqual(true);
-          expect(res.text.includes('data-test-id="error-summary"')).toEqual(true);
-          expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-          orderingPartyController.getCallOffOrderingPartyErrorContext.mockReset();
-        });
-    });
-  });
-
-  describe('GET /organisation/:orderId/supplier', () => {
-    const path = '/organisation/some-order-id/supplier';
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedGetPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()), pathToTest: path, expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedGetPathForUnauthorisedUser({
-        app: request(setUpFakeApp()),
-        pathToTest: path,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should redirect to /organisation/some-order-id/supplier/search', () => request(setUpFakeApp())
-      .get(path)
-      .set('Cookie', [mockAuthorisedCookie])
-      .expect(302)
-      .then((res) => {
-        expect(res.redirect).toEqual(true);
-        expect(res.headers.location).toEqual(`${baseUrl}/organisation/some-order-id/supplier/search`);
-      }));
-  });
-
-  describe('GET /organisation/:orderId/supplier/search', () => {
-    const path = '/organisation/some-order-id/supplier/search';
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedGetPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()), pathToTest: path, expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedGetPathForUnauthorisedUser({
-        app: request(setUpFakeApp()),
-        pathToTest: path,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should return the correct status and text when the user is authorised', () => request(setUpFakeApp())
-      .get(path)
-      .set('Cookie', [mockAuthorisedCookie])
-      .expect(200)
-      .then((res) => {
-        expect(res.text.includes('data-test-id="supplier-search-page"')).toBeTruthy();
-        expect(res.text.includes('data-test-id="error-title"')).toBeFalsy();
-      }));
-  });
-
-  describe('POST /organisation/:orderId/supplier/search', () => {
-    const path = '/organisation/order-1/supplier/search';
-
-    it('should return 403 forbidden if no csrf token is available', () => (
-      testPostPathWithoutCsrf({
-        app: request(setUpFakeApp()), pathToTest: path, mockAuthorisedCookie,
-      })
-    ));
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedPostPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()),
-        csrfPagePath: path,
-        pathToTest: path,
-        mockAuthorisedCookie,
-        expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedPostPathForUnauthorisedUsers({
-        app: request(setUpFakeApp()),
-        csrfPagePath: path,
-        pathToTest: path,
-        mockAuthorisedCookie,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should show the supplier search page with errors if there are validation errors', async () => {
-      supplierSearchController.validateSupplierSearchForm = jest.fn()
-        .mockImplementation(() => ({ success: false }));
-
-      supplierSearchController.getSupplierSearchPageErrorContext = jest.fn()
-        .mockImplementation(() => Promise.resolve({
-          errors: [{ text: 'Supplier name is required', href: '#supplierName' }],
-        }));
-
-      const { cookies, csrfToken } = await getCsrfTokenFromGet({
-        app: request(setUpFakeApp()), csrfPagePath: path, mockAuthorisedCookie,
-      });
-
-      return request(setUpFakeApp())
-        .post(path)
-        .type('form')
-        .set('Cookie', [cookies, mockAuthorisedCookie])
-        .send({ _csrf: csrfToken })
-        .expect(200)
-        .then((res) => {
-          expect(res.text.includes('data-test-id="supplier-search-page"')).toEqual(true);
-          expect(res.text.includes('data-test-id="error-summary"')).toEqual(true);
-          expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-        });
-    });
-
-    it('should redirect to /organisation/some-order-id/supplier/search/select?name=some-supp if no validation errors', async () => {
-      supplierSearchController.validateSupplierSearchForm = jest.fn()
-        .mockImplementation(() => ({ success: true }));
-
-      const { cookies, csrfToken } = await getCsrfTokenFromGet({
-        app: request(setUpFakeApp()), csrfPagePath: path, mockAuthorisedCookie,
-      });
-
-      return request(setUpFakeApp())
-        .post(path)
-        .type('form')
-        .set('Cookie', [cookies, mockAuthorisedCookie])
-        .send({
-          supplierName: 'some-supp',
-          _csrf: csrfToken,
-        })
-        .expect(302)
-        .then((res) => {
-          expect(res.redirect).toEqual(true);
-          expect(res.headers.location).toEqual(`${baseUrl}/organisation/order-1/supplier/search?name=some-supp`);
-        });
-    });
-  });
-
-  describe('GET /organisation/:orderId/supplier/search/select', () => {
-    const path = '/organisation/some-order-id/supplier/search/select?name=some-supp';
-
-    it('should redirect to the login page if the user is not logged in', () => (
-      testAuthorisedGetPathForUnauthenticatedUser({
-        app: request(setUpFakeApp()), pathToTest: path, expectedRedirectPath: 'http://identity-server/login',
-      })
-    ));
-
-    it('should show the error page indicating the user is not authorised if the user is logged in but not authorised', () => (
-      testAuthorisedGetPathForUnauthorisedUser({
-        app: request(setUpFakeApp()),
-        pathToTest: path,
-        mockUnauthorisedCookie,
-        expectedPageId: 'data-test-id="error-title"',
-        expectedPageMessage: 'You are not authorised to view this page',
-      })
-    ));
-
-    it('should show the supplier select if there are no validation errors and suppliers were returned', async () => {
-      supplierSelectController.findSuppliers = jest.fn()
-        .mockImplementation(() => Promise.resolve([
-          { supplierId: 'some-supplier-id', name: 'some-supplier-name' }]));
-
-      supplierSelectController.getSupplierSelectPageContext = jest.fn()
-        .mockImplementation(() => {});
-
-      return request(setUpFakeApp())
-        .get(path)
-        .set('Cookie', [mockAuthorisedCookie])
-        .expect(200)
-        .then((res) => {
-          expect(res.text.includes('data-test-id="error-summary"')).toEqual(false);
-          expect(res.text.includes('data-test-id="error-title"')).toEqual(false);
-          expect(res.text.includes('data-test-id="supplier-select-page"')).toBeTruthy();
-        });
-    });
-
-    it('should show the error page indicating no suppliers found', async () => {
-      supplierSelectController.findSuppliers = jest.fn()
-        .mockImplementation(() => Promise.resolve([]));
-
-      return request(setUpFakeApp())
-        .get(path)
-        .set('Cookie', [mockAuthorisedCookie])
-        .expect(200)
-        .then((res) => {
-          expect(res.text.includes('data-test-id="error-title"')).toEqual(true);
-          expect(res.text.includes(
-            "There are no suppliers that match the search terms you've provided. Try searching again.",
-          )).toEqual(true);
         });
     });
   });
